Real Final Commit

+++ controllers/restaurantController.js
@@ -2,9 +2,13 @@ const Restaurant = require("../models/Restaurant");
 
 exports.createRestaurant = async (req, res) => {
   try {
+    if (req.user.role !== 'admin') {
+      return res.status(403).json({ message: "Only admin can create restaurant" });
+    }
+
     const { name, address, phone, openTime, closeTime } = req.body;
     const restaurant = await Restaurant.create({ name, address, phone, openTime, closeTime });
-    res.json(restaurant);
+    res.status(201).json(restaurant);
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -35,9 +39,14 @@ exports.getRestaurantById = async (req, res) => {
 
 exports.updateRestaurant = async (req, res) => {
   try {
+    if (req.user.role !== 'admin') {
+      return res.status(403).json({ message: "Only admin can update restaurant" });
+    }
+
     const { id } = req.params;
     const updated = await Restaurant.findByIdAndUpdate(id, req.body, { new: true });
     if (!updated) return res.status(404).json({ message: "Restaurant not found" });
+
     res.json(updated);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -46,9 +55,14 @@ exports.updateRestaurant = async (req, res) => {
 
 exports.deleteRestaurant = async (req, res) => {
   try {
+    if (req.user.role !== 'admin') {
+      return res.status(403).json({ message: "Only admin can delete restaurant" });
+    }
+
     const { id } = req.params;
     const deleted = await Restaurant.findByIdAndDelete(id);
     if (!deleted) return res.status(404).json({ message: "Restaurant not found" });
+
     res.json({ message: "Restaurant deleted" });
   } catch (err) {
     res.status(500).json({ error: err.message });
